feat: separate cloud and software development toolkits
