+++ src/cloud-development/utils/index.ts
@@ -1,10 +1,17 @@
-import { ConditionalCheckFailedException } from '@aws-sdk/client-dynamodb';
-import { ConstructMiddleware } from '../construct';
-import { DatabaseError } from '../exceptions';
-import { ConstructTypes } from '../types';
-import getConfig from '../config';
-import { resolve } from 'path';
+
 import * as fs from 'fs';
+import { resolve } from 'path';
+
+import getConfig from '../config';
+import { ConstructMiddleware } from '../construct';
+import { ConstructTypes } from '../types';
+
+
+
+interface GenerateName {
+  prefix?: string;
+  name: string;
+}
 
 export const createFileNameHandler = (dir: string, filePath: string): string => {
   const partialPath = filePath.split('/') || [];
@@ -15,24 +22,6 @@ export const createFileNameHandler = (dir: string, filePath: string): string => 
   partialPath.push(filename);
 
   return `${dir}/${partialPath.join('/')}.${ext}`;
-};
-
-interface GenerateName {
-  prefix?: string;
-  name: string;
-}
-
-export const streamToString = (stream: any) =>
-  new Promise((resolve, reject) => {
-    const chunks: any[] = [];
-    stream.on('data', (chunk: any) => chunks.push(chunk));
-    stream.on('error', reject);
-    stream.on('end', () => resolve(Buffer.concat(chunks).toString('utf8')));
-  });
-
-export const flattenInput = (obj: Record<string, any>) => {
-  const { input, ...rest } = obj;
-  return { ...input, ...rest };
 };
 
 export const constructName = ({ prefix, name }: GenerateName) => {
@@ -68,18 +57,3 @@ export const getConstructs = async <T extends ConstructTypes>(dir: string) => {
   }
   return constructs;
 };
-
-export function CatchDatabaseException(target: any, propertyKey: string, descriptor: PropertyDescriptor) {
-  const method = descriptor.value;
-
-  descriptor.value = async function (...args: any) {
-    try {
-      return await method.apply(this, args);
-    } catch (error) {
-      if (!isConditionalException(error)) throw new DatabaseError(error);
-      throw error;
-    }
-  };
-}
-
-export const isConditionalException = (error: unknown) => error instanceof ConditionalCheckFailedException;
