+++ src/software-development/lib/base-repository.ts
@@ -1,6 +1,7 @@
-import { CustomQueryItemInput, DynamoDbActions } from '../services/dynamodb';
 import { QueryCommandInput } from '@aws-sdk/client-dynamodb';
 import { BackendError } from '../exceptions';
+import { CustomQueryItemInput, DynamoDbActions } from '../services/dynamodb';
+
 
 export type PrimaryKeys = { PK: string; SK?: string };
