+++ src/software-development/lib/index.ts
@@ -3,7 +3,5 @@ export * as Repository from './base-repository';
 export * as Service from './base-service';
 export * as Exceptions from '../exceptions';
 export * as Services from '../services';
-export * as Config from '../config';
-export * as Construct from '../construct';
 export * as Types from '../types';
 export * as Utils from '../utils';
