+++ src/software-development/services/ssm.ts
@@ -1,5 +1,6 @@
 import { SSMClient, GetParameterCommand, GetParameterCommandInput } from '@aws-sdk/client-ssm';
 
+
 export const client = new SSMClient({});
 
 export class SSMActions {
